requirements.txt
# ---- Core Framework ----
fastapi
uvicorn[standard]

# ---- Database & ORM ----
sqlalchemy>=2.0
psycopg2-binary
alembic

# ---- Auth / JWT / Password Hashing ----
python-jose[cryptography]
passlib[bcrypt]

# ---- Environment Settings ----
pydantic-settings

# ---- Email Validation ----
email-validator

# ---- Background Tasks ----
celery
redis

# ---- AI Model (HuggingFace Summarization) ----
transformers
sentencepiece
torch

# ---- Testing ----
httpx
pytest
